fix: 뉴스레터 상세 정보 api 요청 시, undefined로 id를 설정하는 문제 수정
